+++ lib/pages/game/game_view.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+import 'package:game_list/db/model/game.dart';
+
+class GameView extends StatelessWidget {
+  static const routeName = '/gameView';
+
+  @override
+  Widget build(BuildContext context) {
+    final Game game = ModalRoute.of(context).settings.arguments;
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(
+          game.name,
+          overflow: TextOverflow.fade,
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            ClipRRect(
+              borderRadius: BorderRadius.circular(4),
+              child: Image.network(
+                game.coverUrl,
+                width: 128,
+              ),
+            ),
+            Flexible(
+              child: Padding(
+                padding: const EdgeInsets.fromLTRB(8, 0, 0, 0),
+                child: Column(
+                  children: [
+                    Text(
+                      game.name,
+                      style:
+                          TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
+                    ),
+                  ],
+                ),
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}