+++ lib/pages/tabs/games_tab.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-
 import 'package:game_list/db/database.dart';
 import 'package:game_list/db/model/game.dart';
+import 'package:game_list/pages/game/game_view.dart';
 
 class GamesTab extends StatefulWidget {
   @override
@@ -9,20 +9,9 @@ class GamesTab extends StatefulWidget {
 }
 
 class _GamesTabState extends State<GamesTab> {
-  Future<AppDatabase> _database;
-
-  Future<List<Game>> _getAllGame() async {
-    final database = await _database;
-    return database.gameDao.findAllGame();
-  }
-
-  @override
-  void initState() {
-    _database = $FloorAppDatabase
-        .databaseBuilder('game_list.db')
-        .addMigrations([migration1to2]).build();
-    super.initState();
-  }
+  final Future<AppDatabase> _database = $FloorAppDatabase
+      .databaseBuilder('game_list.db')
+      .addMigrations([migration1to2]).build();
 
   @override
   Widget build(BuildContext context) {
@@ -34,17 +23,23 @@ class _GamesTabState extends State<GamesTab> {
             itemCount: snapshot.data.length + 1,
             itemBuilder: (context, index) {
               if (index == 0) {
-                return SizedBox(height: 8);
+                return SizedBox(height: 4);
               } else {
                 return Card(
+                  margin: EdgeInsets.symmetric(horizontal: 8, vertical: 4),
                   child: ListTile(
                     title: Text(snapshot.data[index - 1].name),
-                    subtitle: Text(
-                        'Release: 2020\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Cras ultricies vel tellus sit amet dictum. Donec vel ipsum fringilla sapien molestie eleifend. Aenean blandit aliquam lorem.'),
-                    leading: Image.network(
-                      snapshot.data[index - 1].coverUrl,
+                    subtitle: Text('Release: 2020\nAvg Rating: 9.99'),
+                    leading: CircleAvatar(
+                      backgroundImage: Image.network(
+                        snapshot.data[index - 1].coverUrl,
+                      ).image,
                     ),
-                    onTap: () {},
+                    onTap: () {
+                      Navigator.pushNamed(context, GameView.routeName,
+                          arguments: snapshot.data[index - 1]);
+                    },
+                    isThreeLine: true,
                   ),
                 );
               }
@@ -55,5 +50,10 @@ class _GamesTabState extends State<GamesTab> {
         }
       },
     );
+  }
+
+  Future<List<Game>> _getAllGame() async {
+    final database = await _database;
+    return database.gameDao.findAllGame();
   }
 }