+++ lib/pages/tabs/tabs.dart
@@ -1,0 +1,4 @@
+export 'package:game_list/pages/tabs/games_tab.dart';
+export 'package:game_list/pages/tabs/home_tab.dart';
+export 'package:game_list/pages/tabs/search_tab.dart';
+export 'package:game_list/pages/tabs/settings_tab.dart';