+++ lib/tabs/tabs.dart
@@ -1,4 +1,0 @@
-export 'package:game_list/tabs/games_tab.dart';
-export 'package:game_list/tabs/home_tab.dart';
-export 'package:game_list/tabs/search_tab.dart';
-export 'package:game_list/tabs/settings_tab.dart';